backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database (Async PostgreSQL)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Appwrite Integration
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Validation
pydantic==2.4.0
pydantic-settings==2.0.3
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Additional utilities
python-multipart==0.0.6
passlib[bcrypt]==1.7.4


# Database (SQLAlchemy)
psycopg2

certifi

pytest-cov